✨[Feat] 라우터 이용하여 페이지 생성
