+++ src/component/Finder/Finder.js
@@ -1,0 +1,13 @@
+import React from "react";
+import styled from "styled-components";
+import Layout from "../Layout";
+
+const Finder = () => {
+  return (
+    <Layout>
+      <div>Finder</div>
+    </Layout>
+  );
+};
+
+export default Finder;
